+++ R/create_adam_xml.R
@@ -5,6 +5,7 @@
 #' @param version The version of the define XML to create.  Currently
 #' only 2.0.0 is supported, which is the default.
 #' @returns A vector of XML strings.
+#' @noRd
 create_adam_xml <- function(lst, version) {