+++ tests/testthat/test-write.R
@@ -406,3 +406,29 @@ test_that("write17: write_define() works as expected on CDISC metadata with chec
   expect_equal(file.exists(hf), TRUE)
 
 })
+
+
+
+test_that("write17: write_define() works as expected on perfect CDISC metadata with checks.", {
+
+  fp <- file.path(data_dir, "data/SDTM_METADATA_PERFECT.xls")
+  op <- file.path(base_path, "output")
+
+  df <- file.path(op, "define.sdtm.xml")
+  hf <- file.path(op, "define.sdtm.html")
+
+  if (file.exists(df))
+    file.remove(df)
+
+  if (file.exists(hf))
+    file.remove(hf)
+
+
+  res <- write_define(fp, op, check = TRUE)
+
+  res
+  expect_equal(length(res) == 0, TRUE)
+  expect_equal(file.exists(df), TRUE)
+  expect_equal(file.exists(hf), TRUE)
+
+})